further refactoring, display shutters
