+++ GUI/common.py
@@ -1,0 +1,9 @@
+from PyQt5.QtGui import QColor
+
+RSSI_LOW = QColor("#e74c3c")
+RSSI_MEDIUM = QColor("#ec8826")
+RSSI_GOOD = QColor("#f1c40f")
+RSSI_FULL = QColor("#8bc34a")
+
+ARROW_UP = '▲'
+ARROW_DN = '▼'
