+++ tdmgr.py
@@ -4,7 +4,6 @@ import logging
 import os
 import re
 import sys
-from dataclasses import dataclass
 from json import loads
 
 from PyQt5.QtCore import QDir, QSettings, QSize, Qt, QTimer, QUrl, pyqtSlot
@@ -24,6 +23,7 @@ from PyQt5.QtWidgets import (
 )
 
 from GUI.dialogs import ClearLWTDialog, PrefsDialog
+from models.devices import TasmotaDevicesModel
 
 try:
     from PyQt5.QtWebEngineWidgets import QWebEngineView
@@ -46,7 +46,6 @@ from Util import (
     initial_commands,
     parse_topic,
 )
-from Util.models import TasmotaDevicesModel
 from Util.mqtt import MqttClient
 
 # TODO: rework device export
@@ -55,19 +54,12 @@ __version__ = "0.2.13"
 __tasmota_minimum__ = "6.6.0.17"
 
 
-@dataclass
-class CmdLineSettings:
-    config_location: str
-
-
 class MainWindow(QMainWindow):
-    def __init__(self, cmdline_settings: CmdLineSettings, *args, **kwargs):
+    def __init__(self, *args, **kwargs):
         super(MainWindow, self).__init__(*args, **kwargs)
         self._version = __version__
         self.setWindowIcon(QIcon(":/logo.png"))
         self.setWindowTitle(f"Tasmota Device Manager {self._version}")
-
-        self.cmdline_settings = cmdline_settings
 
         self.menuBar().setNativeMenuBar(False)
 
@@ -645,16 +637,7 @@ def start():
     app.lastWindowClosed.connect(app.quit)
     app.setStyle("Fusion")
 
-    config_location = ''
-    if len(sys.argv) > 1:
-        _config_location = sys.argv[1]
-        if not os.path.exists(_config_location):
-            os.mkdir(_config_location)
-        config_location = _config_location
-
-    cmdline_settings = CmdLineSettings(config_location=config_location)
-
-    MW = MainWindow(cmdline_settings)
+    MW = MainWindow()
     MW.show()
     sys.exit(app.exec_())
